requirements.txt
python-docx==0.8.11
json5==0.9.14 